Abstract out calendar in MainWindow to make it reuseable
